Fix assembler error introduced in e0fec3a
Pass option -M to xa to allow colons to appear in comments. Stop build when XA assembler reports an error. Do not include empty lines in generated header files. Use macro in Automake.am to prevent duplicate code.

By default XA does not allow colons in comments. This resulted in assembler errors due to the added SPDX identifiers. Because of the make recipe not stopping on an assembler error, incomplete and invalid driver headers were created.